migrate Bll and Dal to .NET 6
