+++ WebApiLab.BLL/WebApiLab.BLL.csproj
@@ -1,11 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\WebApiLab.DAL\WebApiLab.DAL.csproj" />
+    <ProjectReference Include="..\WebApiLab.DAL\WebApiLab.Dal.csproj" />
   </ItemGroup>
 
 </Project>
