+++ WebApiLab.DAL/Entities/Category.cs
@@ -1,17 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace WebApiLab.Dal.Entities;
 
-namespace WebApiLab.DAL.Entities
+public class Category
 {
-    public class Category
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
+    public int Id { get; set; }
 
-        public ICollection<Product> Products { get; }
-                                        = new List<Product>();
-    }
+    public string Name { get; set; } = null!;
+
+    public ICollection<Product> Products { get; } = new List<Product>();
 }
