+++ WebApiLab.DAL/Entities/OrderItem.cs
@@ -1,24 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace WebApiLab.Dal.Entities;
 
-namespace WebApiLab.DAL.Entities
+public class OrderItem
 {
-    public class OrderItem
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        public int ProductId { get; set; }
+    public int ProductId { get; set; }
 
-        public Product Product { get; set; }
+    public Product? Product { get; set; }
 
-        public int OrderId { get; set; }
+    public int OrderId { get; set; }
 
-        public Order Order { get; set; }
+    public Order? Order { get; set; }
 
-        public int Quantity { get; set; }
+    public int Quantity { get; set; }
 
-    }
 }
