+++ WebApiLab.DAL/Entities/Product.cs
@@ -1,29 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations.Schema;
 
-namespace WebApiLab.DAL.Entities
+namespace WebApiLab.Dal.Entities;
+
+public class Product
 {
-    public class Product
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        [Column("ProductName")]
-        public string Name { get; set; }
-        public int UnitPrice { get; set; }
+    [Column("ProductName")]
+    public string Name { get; set; } = null!;
 
-        public int CategoryId { get; set; }
-        public Category Category { get; set; }
+    public int UnitPrice { get; set; }
 
-        public ICollection<OrderItem> ProductOrders { get; }
-                                        = new List<OrderItem>();
+    public int CategoryId { get; set; }
+    public Category? Category { get; set; }
 
-        public ICollection<Order> Orders { get; } = new List<Order>();
+    public ICollection<OrderItem> ProductOrders { get; } = new List<OrderItem>();
 
-        public ShipmentRegion ShipmentRegion { get; set; }
-    }
+    public ICollection<Order> Orders { get; } = new List<Order>();
 
+    public ShipmentRegion ShipmentRegion { get; set; }
 }
+
