+++ WebApiLab.DAL/Entities/ShipmentRegion.cs
@@ -1,13 +1,10 @@
-﻿using System;
+﻿namespace WebApiLab.Dal.Entities;
 
-namespace WebApiLab.DAL.Entities
+[Flags]
+public enum ShipmentRegion
 {
-    [Flags]
-    public enum ShipmentRegion
-    {
-        EU = 1,
-        NorthAmerica = 2,
-        Asia = 4,
-        Australia = 8
-    }
+    Eu = 1,
+    NorthAmerica = 2,
+    Asia = 4,
+    Australia = 8,
 }
