+++ WebApiLab.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31005.135
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32210.238
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApiLab.DAL", "WebApiLab.DAL\WebApiLab.DAL.csproj", "{AA817749-03BD-47DF-A847-84856B4B2B82}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApiLab.Dal", "WebApiLab.DAL\WebApiLab.Dal.csproj", "{AA817749-03BD-47DF-A847-84856B4B2B82}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApiLab.BLL", "WebApiLab.BLL\WebApiLab.BLL.csproj", "{8F376D98-9FF5-4241-8FCF-2F2B50CBFB66}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApiLab.Bll", "WebApiLab.BLL\WebApiLab.Bll.csproj", "{8F376D98-9FF5-4241-8FCF-2F2B50CBFB66}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
